Use relative imports and refactored Arc class into shape.py
